留言修正 資料表改動

+++ AlexBlogMVC/Areas/BackEnd/Models/Comment.cs
@@ -7,7 +7,7 @@ public partial class Comment
 {
     public int CommentId { get; set; }
 
-    public int NewsId { get; set; }
+    public long NewsId { get; set; }
 
     public string UserName { get; set; } = null!;
